template: Add region parameter to all APIs
This patch modifies all the roleapi public functions to have a
Region string required parameter which is then used to create a new
context.Context: api.WithRegion(context.Background(), params.Region)
and pass it as the API operation context.

Unit tests have been modified to include a request
assertion which ensures that the region path interpolation is used when
the /platform APIs are being called.

The template.go and params.go files have been split and reorganized into
multiple files each containing a single API context and its parameters.
